feat: WhatsApp MCP (#516)
WhatsApp MCP server support with QR code pairing.

The important change is the addition of Go language to the base MCP
Docker image. Go is required for the `whatsapp-bridge` part of the MCP,
which is responsible for device pair, chat syncs and write tools.

`McpServerConfigSchema` is changed and now oprtionally contains a
`setup` object. It describes the setup which is performed after the
server start. For now, the only setup step available is `log-monitor`
for `whatsapp`.

There is now an individual server restart endpoint. It's used by a
fallback "Restart server" button in QR code dialog when the pairing
process times out.

On the server-side, this setup initializes a QR code log monitor after
starting the server in `sandboxedMcp.start`. This monitor polls logs
periodically, detect the QR code and broadcast the `mcp-setup` event
with the code as a payload. It will also detect when the device is
paired or the pair process timed out or that MCP is connected and no QR
code is needed.

UI is subscribed to the events emitted by QR code log monitor and shows
an indicator, a dot and a badge on the sidebar MCP widget and on the
installed MCP on the connectors page. A click on a dot / icon or a badge
opens the QR code dialog. In case we have just installed an MCP and we
are still on Connectors page, the QR code dialog will automatically
show. It won't pop up in any other cases (e.g. not when the app was
restarted and not when WhatsApp session is expired).

If the pairing process times out, QR code dialog shows a button to
restart a server: to my knowledge, it's the only way to start the
pairing process again.


https://github.com/user-attachments/assets/dea6e58b-4833-44a1-b880-4e545131d84e

Closes #391
PR with MCP catalog update
https://github.com/archestra-ai/website/pull/152

Claim https://github.com/archestra-ai/archestra/issues/391