fix: bulk assign tools 2 agent endpoint (#1045)
A bug in bulk assignment of MCP tools to agents: saving to server takes
a while. There are two sides to this bug:
1. The assignment of the tools themselves took a few seconds because
there were n requests (as many as tools)
2. The refetch of the MCP server and its tools was invoked n times each,
and it also took a while

This adds a bulk MCP tool assignment endpoint, so that there is 1
request to save tools and 1 invalidation of MCP tools.