fix: handle invites to org for existing users (#1273)
Handle sign in of the existing user is invited to a new org. Basically,
this is when you delete a member from /settings/members (note that the
user itself does not get deleted) then invite them anew.

It also extracts the login / register form into a clientside component
so that it can redirect the existing user with an member invitation to
the login.

Related to https://github.com/archestra-ai/archestra/issues/1263