docs: remove default auth secret (#1375)
Latest `better-auth` version requires the secret to be set and to be at
least 32 characters long, and there is no longer a default, so we are
removing it from our docs. When empty, the secret is auto-generated. We
also want to avoid giving some random secret example, because if it's
copy-pasted into deploy script, the security of the instance will be
compromised.