fix: Fix mcp-gateway personal tokens for admins (#1764)
Personal tokens were checking if there is a intersection of user's and
profile's team. It didn't work for admin personal tokens, since admins
are able to access any profie <-> mcp-gw connection in the UI, but were
not able to use their personal token.
Fix it to check profile:admin permission