fix: preserve MCP serviceAccount when re-installing (#1779)
If you install k8s MCP from scratch, it's assigned a special
`archestra-platform-mcp-k8s-operator` service account. But if you edit
the MCP and re-install it, service account is lost. This fix preserves
the service account. It's not shown in Edit MCP Modal, because its name
currently does not matter, it's `archestra-platform-mcp-k8s-operator`
always.